requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.1.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chroma-hnswlib==0.7.2
chromadb==0.4.9
click==8.1.3
coloredlogs==15.0.1
cryptography==41.0.2
dataclasses-json==0.5.14
dnspython==2.3.0
ecdsa==0.18.0
elastic-transport==8.4.0
elasticsearch==8.9.0
email-validator==2.0.0.post2
embedchain==0.0.56
exceptiongroup==1.1.3
fastapi==0.97.0
firebase-admin==6.2.0
Flask==2.3.3
flatbuffers==23.5.26
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.35
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
google-search-results==2.4.2
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
httptools==0.5.0
httpx==0.24.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lanarky==0.7.12
langchain==0.0.279
langchainplus-sdk==0.0.17
langsmith==0.0.33
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
onnxruntime==1.15.1
openai==0.27.10
openapi-schema-pydantic==1.2.4
orjson==3.9.1
overrides==7.4.0
packaging==23.1
pandas==2.1.0
passlib==1.7.4
Pillow==9.5.0
posthog==3.0.2
pq==1.9.1
proto-plus==1.22.3
protobuf==4.23.4
psycopg==3.1.8
psycopg2-binary==2.9.6
pulsar-client==3.3.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.16.1
PyJWT==2.7.0
Pympler==1.0.1
pyparsing==3.1.0
pypdf==3.15.5
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytube==15.0.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
rich==13.5.2
rpds-py==0.10.2
rsa==4.9
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.10
starlette==0.27.0
streamlit==1.26.0
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
twilio==8.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.7
yarl==1.9.1
youtube-transcript-api==0.6.1
zipp==3.16.2
